requirements.txt
moviepy
pillow>=10.0.0
pygemstones
numpy
